// Appraisal-Learning-Process/dict/api/learning/types.ts
/* eslint-disable */
// tslint:disable
/**
 * Appraisal Service
 * Appraisal Service
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface LearningApi {
	postLearningProcessCreate: (request: Api.LearningData | undefined) => Promise<PostLearningProcessCreateResponse>
	deleteLearningProcessDelete: (id: string) => Promise<DeleteLearningProcessDeleteResponse>
	getLearningProcessGet: (id: string) => Promise<GetLearningProcessGetResponse>
	getLearningProcessGetAll: () => Promise<GetLearningProcessGetAllResponse>
	putLearningProcessUpdate: (id: string, request: Api.LearningData | undefined) => Promise<PutLearningProcessUpdateResponse>
}

export type PostLearningProcessCreateResponse = PostLearningProcessCreate201Response | PostLearningProcessCreate401Response | PostLearningProcessCreate404Response | PostLearningProcessCreate422Response

export interface PostLearningProcessCreate201Response {
	status: 201
	body: Api.LearningData
	headers?: never
}

export interface PostLearningProcessCreate401Response {
	status: 401
	body: Api.MessageData
	headers?: never
}

export interface PostLearningProcessCreate404Response {
	status: 404
	body: Api.MessageData
	headers?: never
}

export interface PostLearningProcessCreate422Response {
	status: 422
	body: Api.MessageData
	headers?: never
}

export type DeleteLearningProcessDeleteResponse = DeleteLearningProcessDelete200Response | DeleteLearningProcessDelete401Response | DeleteLearningProcessDelete404Response | DeleteLearningProcessDelete422Response

export interface DeleteLearningProcessDelete200Response {
	status: 200
	body: Api.MessageData
	headers?: never
}

export interface DeleteLearningProcessDelete401Response {
	status: 401
	body: Api.MessageData
	headers?: never
}

export interface DeleteLearningProcessDelete404Response {
	status: 404
	body: Api.MessageData
	headers?: never
}

export interface DeleteLearningProcessDelete422Response {
	status: 422
	body: Api.MessageData
	headers?: never
}

export type GetLearningProcessGetResponse = GetLearningProcessGet200Response | GetLearningProcessGet401Response | GetLearningProcessGet404Response | GetLearningProcessGet422Response

export interface GetLearningProcessGet200Response {
	status: 200
	body: Api.LearningData
	headers?: never
}

export interface GetLearningProcessGet401Response {
	status: 401
	body: Api.MessageData
	headers?: never
}

export interface GetLearningProcessGet404Response {
	status: 404
	body: Api.MessageData
	headers?: never
}

export interface GetLearningProcessGet422Response {
	status: 422
	body: Api.MessageData
	headers?: never
}

export type GetLearningProcessGetAllResponse = GetLearningProcessGetAll200Response | GetLearningProcessGetAll401Response | GetLearningProcessGetAll404Response

export interface GetLearningProcessGetAll200Response {
	status: 200
	body: Api.LearningPagedResultData
	headers?: never
}

export interface GetLearningProcessGetAll401Response {
	status: 401
	body: Api.MessageData
	headers?: never
}

export interface GetLearningProcessGetAll404Response {
	status: 404
	body: Api.MessageData
	headers?: never
}

export type PutLearningProcessUpdateResponse = PutLearningProcessUpdate200Response | PutLearningProcessUpdate401Response | PutLearningProcessUpdate404Response | PutLearningProcessUpdate422Response

export interface PutLearningProcessUpdate200Response {
	status: 200
	body: Api.LearningData
	headers?: never
}

export interface PutLearningProcessUpdate401Response {
	status: 401
	body: Api.MessageData
	headers?: never
}

export interface PutLearningProcessUpdate404Response {
	status: 404
	body: Api.MessageData
	headers?: never
}

export interface PutLearningProcessUpdate422Response {
	status: 422
	body: Api.MessageData
	headers?: never
}